+++ src/Player.h
@@ -11,9 +11,9 @@ public:
 
 	void update() override;
 
-	void mousePressed(int _x, int _y, int _button) override;
-	void mouseDragged(int _x, int _y, int _button) override;
-	void mouseReleased(int _x, int _y, int _button) override;
+	void mousePressed(float _x, float _y, int _button) override;
+	void mouseDragged(float _x, float _y, int _button) override;
+	void mouseReleased(float _x, float _y, int _button) override;
 	void keyPressed(int key) override;
 	void keyReleased(int key) override;
